Extracted display position calculation from the hotspot view model
